requirements.txt
# Python 3.11
docling==2.7.1
spacy==3.7.5
llama-cpp-python==0.3.6
pydantic==2.9.2
pytest==8.3.3


